+++ libreosteoweb/api/renderers.py
@@ -1,7 +1,7 @@
 from rest_framework_csv import renderers
 class PatientCSVRenderer(renderers.CSVRenderer):
 	header = ['family_name', 'original_name', 'first_name', 'birth_date', 'sex',
-	'address_street', 'address_complement', 'address_zipcode', 'address_city', 'phone', 'mobile_phone',
+	'address_street', 'address_complement', 'address_zipcode', 'address_city', 'email', 'phone', 'mobile_phone',
 	'job', 'hobbies', 'smoker', 'laterality', 'important_info', 'current_treatment', 'surgical_history',
 	'medical_history', 'family_history', 'trauma_history','medical_report']