+++ libreosteoweb/migrations/0023_patient_email.py
@@ -1,0 +1,19 @@
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals
+
+from django.db import models, migrations
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ('libreosteoweb', '0022_therapeutsettings_invoice_footer'),
+    ]
+
+    operations = [
+        migrations.AddField(
+            model_name='patient',
+            name='email',
+            field=models.EmailField(max_length=200, verbose_name='Email', blank=True),
+        ),
+    ]